# improve performance of collapseLeadingSlashes