--- a/lib/collapseLeadingSlashes.js
+++ b/lib/collapseLeadingSlashes.js
@@ -8,16 +8,17 @@
  */
 
 function collapseLeadingSlashes (str) {
-  let i = 0
-  for (i; i < str.length; i++) {
+  if (
+    str[0] !== '/' ||
+    str[1] !== '/'
+  ) {
+    return str
+  }
+  for (let i = 2, il = str.length; i < il; ++i) {
     if (str[i] !== '/') {
-      break
+      return str.slice(i - 1)
     }
   }
-
-  return i > 1
-    ? '/' + str.substr(i)
-    : str
 }
 
 module.exports.collapseLeadingSlashes = collapseLeadingSlashes